finish test

+++ src/main/java/card/application/onboarding/model/response/RiskEvaluationResponse.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class RiskEvaluationResponse {
     String emiratesId;
-    int evaluationScore;
+    double evaluationScore;
 }
